1206_update
Three codes and ONET Job corpus data (finished preprocessing)
1. Create dataset(Preprocessing) : This one may be merged with Amjad's resume data code later
2. Semantic_Similarity_Team2: Matching Job-Resume
3. Utils_Team2: Functions

[Job data]
To use Job corpus data, just use below path depends on work space. Use # ['Title' ,'Description_Job']
path = '/home/ubuntu/Project/Data_cleaned'  #Github path = './Data_cleaned'
df_job = pd.read_pickle(os.path.join(path, 'df_Occupation.pkl'))
